separated metrics into several smaller files
